stockmarket_bot/requirements.txt
Django==4.1
djangorestframework==3.14.0
celery==5.3.4
redis==5.0.1
django-celery-beat==2.5.0
psycopg2==2.9.9
coinbase==2.1.0
requests==2.31.0
coinbase-advancedtrade-python==0.1.6
django-filter==23.3
torch==2.1.0
xgboost==2.0.0
ta==0.10.2